+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 - Updated version for public release.
 - Added detailed documentation and examples.
 - Minor improvements and bug fixes.
+
+## 1.0.1 (2024-11-19)
+- Added support for empty match records
+- Changed adding records to match game